Move BackendSparseMatrix to array API instead of utilities
- Move BackendSparseMatrix class from utilities to array/_sparse.py
- Expose BackendSparseMatrix through xp.sparse.BackendSparseMatrix
- Update all imports to use array API instead of utilities
- Remove BackendSparseMatrix from utilities exports
- Centralize sparse matrix functionality in array API
