+++ cuqi/utilities/__init__.py
@@ -17,8 +17,7 @@ from ._utilities import (
     count_within_bounds,
     count_constant_components_1D,
     count_constant_components_2D,
-    piecewise_linear_1D_DoF,
-    BackendSparseMatrix
+    piecewise_linear_1D_DoF
 )
 
 from ._get_python_variable_name import _get_python_variable_name
