+++ cuqi/utilities/_utilities.py
@@ -9,123 +9,6 @@ import copy
 import matplotlib.pyplot as plt
 
 
-class BackendSparseMatrix:
-    """Backend-agnostic wrapper for sparse matrices that provides consistent interface across NumPy and PyTorch backends."""
-    
-    def __init__(self, scipy_sparse_matrix):
-        """Initialize with a scipy sparse matrix."""
-        self._scipy_matrix = scipy_sparse_matrix
-        
-    @property
-    def T(self):
-        """Transpose of the sparse matrix."""
-        return BackendSparseMatrix(self._scipy_matrix.T)
-    
-    @property
-    def shape(self):
-        """Shape of the sparse matrix."""
-        return self._scipy_matrix.shape
-    
-    @property
-    def nnz(self):
-        """Number of non-zero elements in the sparse matrix."""
-        return self._scipy_matrix.nnz
-    
-    def diagonal(self):
-        """Get diagonal elements of the sparse matrix."""
-        diag_elements = self._scipy_matrix.diagonal()
-        # Convert to current backend
-        return xp.array(diag_elements)
-    
-    def __mul__(self, other):
-        """Multiplication with scalars or arrays."""
-        if xp.isscalar(other):
-            # Scalar multiplication returns another BackendSparseMatrix
-            result = other * self._scipy_matrix
-            return BackendSparseMatrix(result)
-        else:
-            # Array multiplication - return the result as an array
-            if xp.get_backend_name() == "pytorch":
-                # Convert other to numpy for scipy operations
-                if hasattr(other, 'cpu'):  # PyTorch tensor
-                    other_np = xp.to_numpy(other)
-                else:
-                    other_np = other
-                result = self._scipy_matrix @ other_np
-                # Convert back to current backend
-                return xp.array(result)
-            else:
-                # NumPy backend
-                result = self._scipy_matrix @ other
-                return result
-    
-    def __rmul__(self, other):
-        """Right multiplication with scalars or arrays."""
-        # Check if other is a scalar (including 0-dimensional arrays)
-        if xp.isscalar(other) or (hasattr(other, 'ndim') and other.ndim == 0):
-            # Scalar multiplication
-            if hasattr(other, 'item'):  # Extract scalar from 0-dim array/tensor
-                other = other.item()
-            result = other * self._scipy_matrix
-            return BackendSparseMatrix(result)
-        else:
-            # Array multiplication from the left
-            if xp.get_backend_name() == "pytorch":
-                # Convert other to numpy for scipy operations
-                if hasattr(other, 'cpu'):  # PyTorch tensor
-                    other_np = xp.to_numpy(other)
-                else:
-                    other_np = other
-                result = other_np @ self._scipy_matrix
-                # Convert back to current backend
-                return xp.array(result)
-            else:
-                # NumPy backend
-                result = other @ self._scipy_matrix
-                return result
-    
-    def __matmul__(self, other):
-        """Matrix multiplication using @ operator."""
-        
-        if xp.get_backend_name() == "pytorch":
-            # Convert other to numpy for scipy operations
-            if hasattr(other, 'cpu'):  # PyTorch tensor
-                other_np = xp.to_numpy(other)
-            else:
-                other_np = other
-            result = self._scipy_matrix @ other_np
-            # Convert back to current backend
-            return xp.array(result)
-        else:
-            # NumPy backend
-            result = self._scipy_matrix @ other
-            return result
-    
-    def __rmatmul__(self, other):
-        """Right matrix multiplication using @ operator."""
-        
-        if xp.get_backend_name() == "pytorch":
-            # Convert other to numpy for scipy operations
-            if hasattr(other, 'cpu'):  # PyTorch tensor
-                other_np = xp.to_numpy(other)
-            else:
-                other_np = other
-            result = other_np @ self._scipy_matrix
-            # Convert back to current backend
-            return xp.array(result)
-        else:
-            # NumPy backend
-            result = other @ self._scipy_matrix
-            return result
-    
-    def toarray(self):
-        """Convert sparse matrix to dense array."""
-        return self._scipy_matrix.toarray()
-    
-    def get_scipy_matrix(self):
-        """Get the underlying scipy sparse matrix for operations that require it."""
-        return self._scipy_matrix
-
 
 def force_ndarray(value,flatten=False):
     if not isinstance(value, xp.ndarray) and value is not None and not xp.issparse(value) and not callable(value):
@@ -279,7 +162,7 @@ def sparse_cholesky(A):
         chol_factor = (LU.L @ (diags(LU.U.diagonal()**0.5))).T
         
         # Return a backend-agnostic sparse matrix wrapper
-        return BackendSparseMatrix(chol_factor)
+        return xp.sparse.BackendSparseMatrix(chol_factor)
     else:
         raise TypeError('The matrix is not positive semi-definite')
